requirements.txt
schedule~=1.2.2
requests~=2.32.3
pandas~=2.2.2
aiogram~=3.7.0
python-dotenv~=1.0.1
